+++ 1BitJam2/Assets/_Scripts/Gameplay/Floor Objects/Loot.cs
@@ -1,8 +1,11 @@
+using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
 public class Loot : MonoBehaviour
 {
-    public float lootValue = 0;
+    public int lootValue = 0;
+
+    
 }
